+++ CHANGES.rst
@@ -1,6 +1,9 @@
 0.4.0
 =====
 
+- Changed redirect path after reply to last 5 posts instead of full topic.
+- Changed link target for topic title to recent posts instead of full topic.
+- Fixed stylings for error page when post wasn't successful.
 - Added top-right and bottom-right buttons for jumping to header and footer.
 - Added auto-link functionality.